--- rx-wecom-sdk/src/main/java/cn/felord/reactive/callbacks/CallbackAsyncConsumer.java
@@ -1,91 +1,0 @@
-/*
- *  Copyright (c) 2023. felord.cn
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *       https://www.apache.org/licenses/LICENSE-2.0
- *  Website:
- *       https://felord.cn
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-
-package cn.felord.reactive.callbacks;
-
-import cn.felord.domain.callback.CallbackEventBody;
-
-import java.util.concurrent.ArrayBlockingQueue;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.ThreadPoolExecutor;
-import java.util.concurrent.TimeUnit;
-import java.util.function.Consumer;
-
-/**
- * The interface Callback consumer.
- *
- * @author felord.cn
- * @since 1.0.14.RELEASE
- */
-public final class CallbackAsyncConsumer {
-    private final ExecutorService executor;
-    private final Consumer<CallbackEventBody> eventBodyConsumer;
-
-
-    /**
-     * Instantiates a new Callback async consumer.
-     *
-     * @param eventBodyConsumer the event body consumer
-     */
-    public CallbackAsyncConsumer(Consumer<CallbackEventBody> eventBodyConsumer) {
-        this(10, 100, 90L, 500, eventBodyConsumer);
-    }
-
-    /**
-     * Instantiates a new Callback async consumer.
-     *
-     * @param corePoolSize      the core pool size
-     * @param maximumPoolSize   the maximum pool size
-     * @param keepAliveSecond   the keep alive second
-     * @param capacity          the capacity
-     * @param eventBodyConsumer the event body consumer
-     */
-    public CallbackAsyncConsumer(int corePoolSize,
-                                 int maximumPoolSize,
-                                 long keepAliveSecond,
-                                 int capacity,
-                                 Consumer<CallbackEventBody> eventBodyConsumer) {
-
-        this(new ThreadPoolExecutor(corePoolSize, maximumPoolSize,
-                keepAliveSecond, TimeUnit.SECONDS,
-                new ArrayBlockingQueue<>(capacity),
-                new CustomizableThreadFactory("WECOM-CALLBACK-THREAD-POOL-")), eventBodyConsumer);
-
-    }
-
-
-    /**
-     * Instantiates a new Callback async consumer.
-     *
-     * @param executor          the executor
-     * @param eventBodyConsumer the event body consumer
-     */
-    public CallbackAsyncConsumer(ExecutorService executor, Consumer<CallbackEventBody> eventBodyConsumer) {
-        this.executor = executor;
-        this.eventBodyConsumer = eventBodyConsumer;
-    }
-
-    /**
-     * Action.
-     *
-     * @param eventBody the event body
-     */
-    public void asyncAction(CallbackEventBody eventBody) {
-        Thread thread = new Thread(() -> eventBodyConsumer.accept(eventBody));
-        executor.execute(thread);
-    }
-
-
-}
