+++ wecom-objects/src/main/java/cn/felord/callbacks/CallbackSettings.java
