+++ wecom-sdk/src/main/java/cn/felord/TokenInterceptor.java
@@ -1,3 +1,18 @@
+/*
+ *  Copyright (c) 2023. felord.cn
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *       https://www.apache.org/licenses/LICENSE-2.0
+ *  Website:
+ *       https://felord.cn
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
 package cn.felord;
 
 import cn.felord.api.AccessTokenApi;
@@ -9,8 +24,6 @@ import okhttp3.*;
 import org.jetbrains.annotations.NotNull;
 
 import java.io.IOException;
-import java.util.Arrays;
-import java.util.List;
 import java.util.Objects;
 
 /**
@@ -22,9 +35,6 @@ import java.util.Objects;
 public class TokenInterceptor implements Interceptor {
     public static final Integer INVALID_ACCESS_TOKEN = 42001;
     public static final ObjectMapper MAPPER = JacksonObjectMapperFactory.create();
-    private static final List<String> WHITE_LIST = Arrays.asList("/cgi-bin/webhook/send",
-            "/cgi-bin/webhook/upload_media",
-            "/cgi-bin/gettoken");
     private final TokenApi tokenApi;
     private final String tokenParam;
