--- wecom-sdk/src/main/java/cn/felord/callbacks/ByteCollector.java
@@ -1,40 +1,0 @@
-/*
- * Copyright (c) 2023. felord.cn
- *   Licensed under the Apache License, Version 2.0 (the "License");
- *   you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *      https://www.apache.org/licenses/LICENSE-2.0
- * Website:
- *      https://felord.cn
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package cn.felord.callbacks;
-
-class ByteCollector {
-    private byte[] byteContainer;
-
-    public byte[] toBytes() {
-        return byteContainer;
-    }
-
-    public ByteCollector addBytes(byte[] bytes) {
-        if (byteContainer == null) {
-            this.byteContainer = bytes;
-        } else {
-            byte[] newByte = new byte[byteContainer.length + bytes.length];
-            System.arraycopy(bytes, 0, newByte, 0, bytes.length);
-            System.arraycopy(byteContainer, 0, newByte, 0, byteContainer.length);
-            this.byteContainer = newByte;
-        }
-        return this;
-    }
-
-    public int size() {
-        return byteContainer == null ? 0 : byteContainer.length;
-    }
-}
